refactor: rename view
- now distribution list has its own dedicated view
