+++ src/views/distribution-list/distribution-lists-view.test.tsx
@@ -14,6 +14,7 @@ import { HttpResponse } from 'msw';
 import { Link, Route } from 'react-router-dom';
 
 import { DistributionListsView } from './distribution-lists-view';
+import GroupsAppView from './DistributionListAppView';
 import { screen, setupTest, within } from '../../carbonio-ui-commons/test/test-setup';
 import { ROUTES, ROUTES_INTERNAL_PARAMS } from '../../constants';
 import {
@@ -38,7 +39,6 @@ import {
 	generateDistributionList,
 	generateDistributionLists
 } from '../../tests/utils';
-import GroupsAppView from '../contact-groups/GroupsAppView';
 
 describe('Distribution Lists View', () => {
 	it('should show the list of distribution lists', async () => {
